requirements.txt
google-generativeai 
langchain-community 
sentence-transformers 
faiss-cpu 
pypdf